Move default implementation to protocol extension.
